requirements.txt
imap_tools==0.50.2
pycurl==7.43.0.6
lxml==4.5.0
python-dotenv==0.19.2
